Android 16.0.0 release 4

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCaS5huQAKCRDorT+BmrEO
# eK5QAJ46InBs2dUOIR4NkSiDFJks9WHckgCeKEzctcWXH2+IUijtvoynZrtv2tY=
# =SeE5
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue Dec  2 05:49:13 2025 EET
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [ultimate]

* tag 'android-16.0.0_r4': (1338 commits)
  Import translations. DO NOT MERGE ANYWHERE
  RFCOMM MUX connection request must not start bonding
  Import translations. DO NOT MERGE ANYWHERE
  Fix Subrate Disablement Peripheral Latency restore
  Add flag hid_connection_timeout_in_jni_thread
  Fix scan result callback description in btm_ble_gap.cc
  pandora: Add gRPC APIs to support MAP client
  Update usage description for dial() API
  Do not require BLE_ON for duration of the test
  Remove flag auto_transport_pairing flag and related logic
  Bluetooth::satellite Update bug reference for proper API
  Only broadcast to concerned user
  BluetoothGatt Use annotation for NonNull
  25Q4 OWNER freeze window
  floss: Fix stop_watch missing <mutex> header
  Make PBAP server provide high resolution photos
  Check if `ScanController` is available before calling in AdapterSuspend
  Add modern Java syntax for generic types
  Refactor the parsing tool for dumping ASHA audio data from btsnoop
  snoop_logger_file: create SnoopLoggerFile to handle file operations
  ...

 Conflicts:
	system/bta/le_audio/client.cc
	system/bta/le_audio/codec_manager.cc
	system/bta/le_audio/codec_manager_test.cc
	system/bta/le_audio/devices.cc
	system/bta/le_audio/le_audio_client_test.cc
	system/bta/le_audio/state_machine.cc

+++ android/app/tests/unit/src/com/android/bluetooth/map/BluetoothMapContentObserverTest.java
@@ -16,7 +16,6 @@
 
 package com.android.bluetooth.map;
 
-import static com.android.bluetooth.TestUtils.MockitoRule;
 import static com.android.bluetooth.TestUtils.mockGetSystemService;
 
 import static com.google.common.truth.Truth.assertThat;
@@ -54,13 +53,14 @@ import android.telephony.TelephonyManager;
 import android.test.mock.MockContentProvider;
 import android.test.mock.MockContentResolver;
 
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.MediumTest;
-import androidx.test.runner.AndroidJUnit4;
 
 import com.android.bluetooth.BluetoothMethodProxy;
 import com.android.bluetooth.map.BluetoothMapContract.MessageColumns;
 import com.android.bluetooth.map.BluetoothMapUtils.TYPE;
 import com.android.obex.ResponseCodes;
+import com.android.tests.bluetooth.MockitoRule;
 
 import com.google.android.mms.pdu.PduHeaders;
 
@@ -2021,9 +2021,9 @@ public class BluetoothMapContentObserverTest {
     @Test
     public void handleMmsSendIntent_withInvalidHandle() {
         when(mClient.isConnected()).thenReturn(true);
-        doReturn((long) -1)
+        doReturn(-1L)
                 .when(mIntent)
-                .getLongExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_HANDLE, -1);
+                .getLongExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_HANDLE, -1L);
 
         assertThat(mObserver.handleMmsSendIntent(mContext, mIntent)).isTrue();
     }
@@ -2074,9 +2074,9 @@ public class BluetoothMapContentObserverTest {
         doReturn(1)
                 .when(mIntent)
                 .getIntExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_TRANSPARENT, 0);
-        doReturn((long) -1)
+        doReturn(-1L)
                 .when(mIntent)
-                .getLongExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_HANDLE, -1);
+                .getLongExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_HANDLE, -1L);
         // This mock sets type to MMS
         doReturn(4)
                 .when(mIntent)
@@ -2115,9 +2115,9 @@ public class BluetoothMapContentObserverTest {
         doReturn(1)
                 .when(mIntent)
                 .getIntExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_TRANSPARENT, 0);
-        doReturn((long) -1)
+        doReturn(-1L)
                 .when(mIntent)
-                .getLongExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_HANDLE, -1);
+                .getLongExtra(BluetoothMapContentObserver.EXTRA_MESSAGE_SENT_HANDLE, -1L);
 
         mObserver.actionMmsSent(mContext, mIntent, 1, mmsMsgList);
